Refactor mock server and set up root endpoint
