+++ api/src/champion-mastery/controller.ts
@@ -1,0 +1,8 @@
+import { API } from 'lambda-api';
+import { getChampionMastery } from './service';
+
+export const championMasteryApi = (api: API) => {
+  api.get('/', async () => {
+    return getChampionMastery();
+  });
+};
