+++ api/src/champion-mastery/service.ts
@@ -1,0 +1,3 @@
+export const getChampionMastery = () => {
+  return 'Hello';
+};
