--- api/src/handler.ts
@@ -1,3 +1,0 @@
-export const handler = () => {
-  return 'Hello';
-};
